==> 01B_text_preprocessing_for_label.ipynb <==
import pandas as pd
import numpy as np
import re
import string

from sklearn.feature_extraction.text import TfidfVectorizer
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
import nltk

nltk.download('punkt')
nltk.download('stopwords')


data = pd.read_csv("data/processed/combined_fake_news_dataset.csv") 
data

def clean_text(text):
    text = str(text).lower()                        # lowercase
    text = re.sub(r"http\S+", "", text)             # remove URLs
    text = re.sub(r"\d+", "", text)                 # remove digits
    text = re.sub(r"[^\w\s]", "", text)             # remove punctuation
    return text

data["text_clean"] = data["text"].apply(clean_text)


import tensorflow as tf
print(tf.__version__)



  import spacy

# Load the small English model
nlp = spacy.load("en_core_web_sm")   


def spacy_tokenize_and_remove_stopwords(text):
    doc = nlp(text)
    tokens = [token.text for token in doc if not token.is_stop and not token.is_punct]
    return " ".join(tokens)

data["text_tokens"] = data["text_clean"].apply(spacy_tokenize_and_remove_stopwords)


tfidf = TfidfVectorizer(max_features=5000)
X = tfidf.fit_transform(data["text_tokens"])



import joblib
from sklearn.preprocessing import LabelEncoder

# Map labels manually if needed (though LabelEncoder does this automatically)
y = data["label"].map({"real": 0, "fake": 1}) 

# Create and fit LabelEncoder
le_label = LabelEncoder()
le_label.fit(["real", "fake"])  # Fitting the encoder

# ✅ Save the fitted encoder
joblib.dump(le_label, "models/source_label_encoder.pkl")


pd.DataFrame(X.toarray()).to_csv("data/processed/vectorized_text.csv", index=False)
pd.DataFrame(y).to_csv("data/processed/labels.csv", index=False)
