==> 01c_text_preprocessing_for_type.ipynb <==
import pandas as pd

df = pd.read_csv("data/processed/combined_news_type_dataset.csv")
df.head()


print(df['subject'].value_counts())



# Define the mapping dictionary
subject_map = {
    "politicsNews": "politics",
    "politics": "politics",
    "Government News": "politics",
    "left-news": "politics",
    "US_News": "national",
    "Middle-east": "world",
    "worldnews": "world",
    "News": "general",
    "entertainment": "entertainment"
}

# Apply the mapping
df["subject_mapped"] = df["subject"].map(subject_map)

# Check new class distribution
print(df["subject_mapped"].value_counts())


import spacy
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
import pandas as pd


nlp = spacy.load("en_core_web_sm")

def spacy_tokenizer(text):
    doc = nlp(text.lower())
    tokens = [token.lemma_ for token in doc if not token.is_stop and not token.is_punct and not token.is_space]
    return " ".join(tokens)

df["text_clean"] = df["text"].astype(str).apply(spacy_tokenizer)


tfidf = TfidfVectorizer(max_features=5000)
X = tfidf.fit_transform(df["text_clean"])


le = LabelEncoder()
y = le.fit_transform(df["subject_mapped"])


X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size=0.3, 
                                                    random_state=42, 
                                                    stratify=y)


import joblib

joblib.dump(tfidf, "models/spacy_tfidf_vectorizer.pkl")
joblib.dump(le, "models/label_encoder.pkl")

pd.DataFrame(X_train.toarray()).to_csv("data/processed/X_train.csv", index=False)
pd.DataFrame(X_test.toarray()).to_csv("data/processed/X_test.csv", index=False)
pd.DataFrame({"label": y_train}).to_csv("data/processed/y_train.csv", index=False)
pd.DataFrame({"label": y_test}).to_csv("data/processed/y_test.csv", index=False)
